Fix comparison error

+++ scene/main/node.cpp
@@ -2049,7 +2049,7 @@ TypedArray<Node> Node::find_children(const String &p_pattern, const String &p_ty
 				while (current_node != this) {
 					const LocalVector<Node *> &siblings = current_node->data.parent->data.children_cache;
 
-					if (current_node->data.index + 1 < siblings.size()) {
+					if (current_node->data.index + 1 < (int)siblings.size()) {
 						// Go to next sibling
 						current_node = siblings[current_node->data.index + 1];
 						break;
